BUG: Install SlicerRadiomicsCLI script on windows
On Unix, SlicerRadiomicsCLI requires the `SlicerRadiomicsCLI.xml` and `SlicerRadiomicsCLI` files (description and execution). On windows an additional batch script file (`SlicerRadiomicsCLI.bat`) is required, which executes `SlicerRadiomicsCLI` using the Slicer Python interpreter. However, the installation on windows installed the batch script instead of the execution file, not in addition to. Update the SlicerRadiomicsAddCLI function to correct this.