BUG: Fix permissions of CLI script
To enable detection and execution of the CLI script, it must have executable permissions. Update the cmake file to assign this permission for the user to the script file in the build directory (enabling the user to use the script without having to install it), but only when the system is not WIN32. In addition, handle installation of the script separately and install it as PROGRAM, which will assign to correct permissions to the packaged file.