ENH: Allow reusing of the CLI node in one extraction
Reuse the CLI node when extracting multiple Labels/ROIs/Segmentations.
Remove usage of a temporary table, port output to stdout instead and intercept and parse this message into the output table.

Additionally, rewrite the tests to work with the new interface. Add a while loop that waits for the extraction to finish.