BUG: Fix interpreter name
When running the script from a build-tree of slicer (compiled), the python interpreter is named `python`, however, in an installed/packaged tree, the python interpreter is named `python-real`.

Update the batch script (for windows) to first try to use `python-real` and if not found, use `python` instead.

Similarly, add a shell script (for Linux/Mac) to also first try to use `python-real` and upon fail, use `python` instead.

Rename the SlicerRadiomicsCLI python script to `SlicerRadiomicsCLIScript`, as `SlicerRadiomicsCLI` is now used for the shell script file.

Fixes #40