BUG: Fix incompatibilyt due to commandline refactoring in pyradiomics
Recently a [PR](https://github.com/Radiomics/pyradiomics/pull/347) was merged in PyRadiomics, refactoring the commandline interface. Update the SlicerRadiomics module to work with the new functionality.

Additionally, fix a bug when requesting the XML from the SlicerRadiomicsScript (due to a name change, it is now necessary to cut off "Script").