fix value error (and missing data) in table
The old implementation was indexing beyond the end of the
rows because it does not add rows for skipped keys, but did
use the key index when setting cell values.

Importantly this led to missing data (typically the last two rows
due to the Image and Mask keys.

In old VTK python this was silently accepted and might cause
memory corruption.  But VTK's python wrapping now does bounds
checks, so the old implementation led to ValueError exceptions.

This fixed code now tracks the skipped keys and adjusts the row
index accordingly.