ENH: Update CLI definition
Remove use .bat and .sh wrapper script, and instead define the CLI entry point as a python script.
This makes use of the new Slicer functionality described in https://github.com/Slicer/Slicer/pull/894 and implemented in r27143-r27145.

This fixes #43.

Additionally, check if a `--label <label>` argument is present, and if so, update it to the new-style: `--setting=label:<label>`. Parameter `--label` is deprecated by PR https://github.com/Radiomics/pyradiomics/pull/347.