BUG: Revert Fix PyWavelets version to <= 1.0.0
Specifying the version in the no-binary part of the pip call breaks this argument, allowing slicer to install from the wheel. This in turn causes the DLL to fail to load, as a different compiler is used on Windows.
Revert the change here. The version is also fixed to <=1.0.0 in the PyRadiomics repository, and will therefore install the correct PyWavelets, including compiling it using Slicer's build setup.