BUG: Fix some segment missing in output
When input regions were provided as a segmentation node, segments were missed when segment name was not the same as the segment ID (e.g., the user renamed the segment).

The problem was that ExportSegmentsToLabelmapNode requires segment ID, but it was provided segment name instead. Fixed the issue by providing segment ID.